Create a parent class for shopping tests
